+++ apps/demo/src/app/app.module.ts
@@ -3,18 +3,26 @@ import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
 import {
-  LoadMermaidInjectorToken,
   MarkdownModule,
-  LoadMarkdownWorkerInjectorToken,
-  GetWorkerPayload
 } from '@gewd/markdown';
 import { FormsModule } from '@angular/forms';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatTabsModule } from '@angular/material/tabs';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { GetWorkerPayload, LoadMarkdownWorkerInjectorToken, LoadMermaidInjectorToken } from '@gewd/markdown/contracts';
 
-const marked =  () =>  new Worker('./markdown.worker.ts', { type: 'module' });
+const marked = () => new Worker('./markdown.worker.ts', { type: 'module' });
+
+console.info(LoadMarkdownWorkerInjectorToken, LoadMermaidInjectorToken);
 
 @NgModule({
   declarations: [AppComponent],
-  imports: [BrowserModule, MarkdownModule, FormsModule],
+  imports: [
+    BrowserModule,
+    MarkdownModule, FormsModule,
+    MatExpansionModule, MatTabsModule,
+    BrowserAnimationsModule
+  ],
   providers: [
     {
       provide: LoadMarkdownWorkerInjectorToken,
@@ -29,4 +37,5 @@ const marked =  () =>  new Worker('./markdown.worker.ts', { type: 'module' });
   ],
   bootstrap: [AppComponent]
 })
-export class AppModule {}
+export class AppModule {
+}