+++ libs/markdown/module/public_api.ts
@@ -1,0 +1,2 @@
+export * from './markdown.module';
+export * from './render-markdown/render-markdown.component';