+++ libs/markdown/src/index.ts
@@ -1,6 +1,0 @@
-
-export * from './lib/markdown.module';
-export * from './lib/render-markdown/render-markdown.component';
-export * from './lib/worker.service';
-
-export * from './lib/contracts';