+++ libs/markdown/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.2.0] - 2019-12-18
+
+- :sparkles: Emoji Support inside Markdown-Worker
+  Can be also used alone: 
+  - `@gewd/markdown/emoji-map` 
+     - `EMOJI_MAP`
+     - `colonToUnicode(colonName)`
+- :zap: more entry points to optimize builds
+  - `@gewd/markdown/utils`
+  - `@gewd/markdown/service`
+
 ## [0.1.0] - 2019-12-16
 
 Start of Changelog